% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alfonsino}
\alias{alfonsino}
\title{Example alfonsino data}
\format{
The dataset loads three objects:
\describe{
\item{\code{lhpar_alfonsino}}{An \code{FLPar} object with life-history parameters.}
\item{\code{lfd_alfonsino}}{An \code{FLQuants} object with simulated length-frequency data.}
\item{\code{stklen_alfonsino}}{An \code{FLStockLen} example object.}
}
}
\source{
FLife simulation
}
\usage{
data(alfonsino)
}
\description{
Example data objects for alfonsino, including life-history parameters,
length-frequency data, and an example \code{FLStockLen} object.
}
\author{
Henning Winker
}
