% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicccbrp.R
\name{brp_flqs_flicc}
\alias{brp_flqs_flicc}
\title{Extract BRP plotting quantities from an FLStockLen BRP object}
\usage{
brp_flqs_flicc(stk)
}
\arguments{
\item{stk}{An object returned by \code{brp_stklen_flicc()}.}
}
\value{
An \code{FLQuants} object with \code{F}, \code{Rec}, \code{Catch},
and \code{SSB}.
}
\description{
Extract BRP plotting quantities from an FLStockLen BRP object
}
