% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccc.checks.R
\name{compare_zlcc_flicc}
\alias{compare_zlcc_flicc}
\title{Compare empirical LCCC Z with FLicc model-implied Z}
\usage{
compare_zlcc_flicc(
  fit,
  lfd,
  lc = NULL,
  empirical_lc_args = list(),
  max_rel_linf = 0.999,
  min_n = 3,
  weight = c("lfd", "equal")
)
}
\arguments{
\item{fit}{A fitted FLicc object containing \code{fit$stklen} and
\code{fit$report$Fap}.}

\item{lfd}{Gear-specific FLQuant length-frequency data, e.g.
\code{lfd_alfonsino$Gillnet}.}

\item{lc}{Optional lower length cutoff for LCCC. If \code{NULL}, estimated
using \code{empirical_lc()}.}

\item{empirical_lc_args}{Optional list of arguments passed to
\code{empirical_lc()}.}

\item{max_rel_linf}{Maximum allowed length relative to Linf.}

\item{min_n}{Minimum number of observations for each LCCC regression.}
}
\value{
A list containing empirical Z, FLicc Z, merged comparison table,
regression data, lc and empirical_lc.
}
\description{
Compares empirical total mortality estimated from length-converted
catch-curve regressions with the FLicc model-implied total mortality
proxy M + Fap. The empirical Z is gear-specific if a gear-specific
length-frequency object is supplied.
}
\examples{
\dontrun{
data(alfonsino)

lfds <- lfd_alfonsino
lfd <- lfds$Gillnet
lhpar <- lhpar_alfonsino

stklen <- stocklen(lfds, lhpar)
fit <- fiticc(
  lfds, stklen,
  sel_fun = c("dsnormal", "logistic"),
  catch_by_gear = c(0.7, 0.3)
)

stkfit <- flicc_stklen(fit)

cmp <- compare_zlcc_flicc(fit, lfd =lfd)
cmp$table
plot_compare_zlcc_flicc(cmp, quantity = "Z")+ylim(0,NA)
}

}
