% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStockLen.R
\name{compute_catch.n}
\alias{compute_catch.n}
\title{Compute catch numbers from stock numbers and mortalities}
\usage{
compute_catch.n(stock.n, harvest, m)
}
\arguments{
\item{stock.n}{An \code{FLQuant} of stock numbers.}

\item{harvest}{An \code{FLQuant} of fishing mortality.}

\item{m}{An \code{FLQuant} of natural mortality.}
}
\value{
An \code{FLQuant} of catch numbers.
}
\description{
Computes catch-at-length or catch-at-age in numbers using the Baranov
catch equation.
}
