% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStockLen.R
\name{fapl}
\alias{fapl}
\title{Fully selected fishing mortality from an FLStockLen BRP object}
\usage{
fapl(stk)
}
\arguments{
\item{stk}{An object returned by \code{brp_stklen_flicc()}.}
}
\value{
An \code{FLQuant} over the BRP year dimension.
}
\description{
Fully selected fishing mortality from an FLStockLen BRP object
}
