% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStockLen.R
\name{fbarl}
\alias{fbarl}
\title{Mean fishing mortality over a length range}
\usage{
fbarl(stk, lrange = NULL)
}
\arguments{
\item{stk}{An \code{FLStockLen} object.}

\item{lrange}{Optional numeric vector of length 2 giving min and max length.}
}
\value{
An \code{FLQuant}.
}
\description{
Mean fishing mortality over a length range
}
