% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcflicc.R
\name{fill_report_iter}
\alias{fill_report_iter}
\title{Fill one iteration of a propagated FLicc report}
\usage{
fill_report_iter(mcrep, report_i, i)
}
\arguments{
\item{mcrep}{A propagated report from \code{prop_report()}.}

\item{report_i}{A report list from one Monte Carlo fit.}

\item{i}{Iteration number.}
}
\value{
Updated Monte Carlo report.
}
\description{
Fill one iteration of a propagated FLicc report
}
