% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcflicc.R
\name{fishlife_corr}
\alias{fishlife_corr}
\title{Default FishLife-style life-history correlation matrix}
\usage{
fishlife_corr()
}
\value{
A numeric correlation matrix.
}
\description{
Returns a default correlation matrix for drawing correlated life-history
parameters in \code{mc_flicc()}.
}
