% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fliccit.R
\name{gtg_inputs_flicc}
\alias{gtg_inputs_flicc}
\title{Extract FLicc population model name}
\usage{
gtg_inputs_flicc(fit)
}
\arguments{
\item{fit}{A fitted \code{"flicc_tmb_fit"} object.}
}
\value{
Character string, one of \code{"gamma"} or \code{"gtg"}.
}
\description{
Extract FLicc population model name
}
