% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfd_alfonsino}
\alias{lfd_alfonsino}
\title{lfd_alfonsino}
\format{
an \code{FLQuants} object
}
\source{
FLife simulation
}
\usage{
lfd_alphonsino
}
\description{
Simulated LFD data from a trawl and gillnet fishery for Alphonsino
}
\author{
Henning Winker
}
