% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LLflicc}
\alias{LLflicc}
\title{Extract log-likelihood from a fitted FLicc model}
\usage{
LLflicc(object, ...)
}
\arguments{
\item{object}{A fitted \code{"flicc_tmb_fit"} object.}

\item{...}{Additional arguments, ignored.}
}
\value{
An object of class \code{logLik}.
}
\description{
Returns the fitted log-likelihood from a \code{"flicc_tmb_fit"} object as a
\code{logLik}-class object. The value is computed as the negative of the
TMB objective function at the optimum.
}
\details{
The returned object includes \code{df}, taken as the number of estimated
parameters in \code{object$opt$par}, and \code{nobs}, taken as the number of
rows in \code{as.data.frame(object$report$obslen)}.

This helper assumes that \code{object$opt$objective} is the negative
log-likelihood returned by TMB. If penalties or priors were included in the
objective function, the returned value should be interpreted as the
corresponding penalized log-likelihood or objective-based criterion rather
than a pure data likelihood.
}
\examples{
\dontrun{
LL <- LLflicc(fit)
LL
AIC(LL)
BIC(LL)
}

}
