% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m_gislason}
\alias{m_gislason}
\title{Gislason natural mortality at length}
\usage{
m_gislason(len, linf, k)
}
\arguments{
\item{len}{Numeric vector of lengths, usually bin midpoints.}

\item{linf}{Asymptotic length.}

\item{k}{Von Bertalanffy growth coefficient.}
}
\value{
Numeric vector of natural mortality values at length.
}
\description{
Computes natural mortality at length using the Gislason et al. length-based
empirical formula.
}
