% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccc.checks.R
\name{plot_compare_zlcc_flicc}
\alias{plot_compare_zlcc_flicc}
\title{Plot empirical LCCC Z/F against FLicc model-implied Z/F}
\usage{
plot_compare_zlcc_flicc(x, quantity = c("Z", "F"))
}
\arguments{
\item{x}{Output from \code{compare_zlcc_flicc()}.}

\item{quantity}{Either \code{"Z"} or \code{"F"}.}
}
\value{
A ggplot object.
}
\description{
Plot empirical LCCC Z/F against FLicc model-implied Z/F
}
