% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicc_plots.R
\name{plot_mat}
\alias{plot_mat}
\title{Plot maturity-at-length curves}
\usage{
plot_mat(object, nyears = 1, len_by = 5, colours = NULL, line_width = 0.8)
}
\arguments{
\item{object}{An \code{FLStockLen} or \code{FLStocks} object.}

\item{nyears}{Integer number of most recent years over which to average
\code{mat()}. Default is \code{1}.}

\item{len_by}{Numeric spacing for length-axis tick marks. Default is
\code{5}.}

\item{colours}{Optional named character vector of colours. If \code{NULL},
colours are assigned automatically using the same generic palette logic as
other FLicc plotting functions.}

\item{line_width}{Numeric line width for maturity curves. Default is
\code{0.8}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots mean maturity-at-length from an \code{FLStockLen} or
\code{FLStocks} object. If multiple years are present, the plot uses the
mean over the last \code{nyears}.
}
\details{
If a single \code{FLStockLen} object is supplied, it is internally coerced
to an \code{FLStocks} object named \code{mat_l}.
}
\examples{
\dontrun{
plot_mat(stklen)
plot_mat(stklen, nyears = 3)
plot_mat(stklen, len_by = 5)
plot_mat(stklen, colours = c(mat_l = "black"))
}

}
