% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccc.checks.R
\name{plot_zlcc}
\alias{plot_zlcc}
\title{Plot length-converted catch-curve regressions}
\usage{
plot_zlcc(
  lfd,
  lhpar,
  lc = NULL,
  empirical_lc_args = list(),
  min_n = 3,
  max_rel_linf = 0.999,
  facet = TRUE,
  show_points = TRUE,
  show_lm = TRUE,
  show_eq = TRUE,
  ncol = NULL,
  return_data = FALSE
)
}
\arguments{
\item{lfd}{An FLQuant with length in the first dimension.}

\item{lhpar}{Life-history parameters containing at least \code{linf}
and \code{k}.}

\item{lc}{Optional lower length cutoff for the descending limb. If
\code{NULL}, this is estimated with \code{empirical_lc()}.}

\item{empirical_lc_args}{Optional list of arguments passed to
\code{empirical_lc()}.}

\item{min_n}{Minimum number of observations required to fit a regression.}

\item{max_rel_linf}{Maximum allowed length relative to Linf.}

\item{facet}{Logical. Should regressions be shown in facets?}

\item{show_points}{Logical. Show observed regression points.}

\item{show_lm}{Logical. Show fitted linear regression line.}

\item{show_eq}{Logical. Add labels with Z, R-squared and n.}

\item{ncol}{Optional number of facet columns.}

\item{return_data}{Logical. If TRUE, return a list with the plot,
regression data and fitted Z table.}
}
\value{
A ggplot object, or a list if \code{return_data = TRUE}.
}
\description{
Produces diagnostic plots for the linear regressions used to estimate
total mortality, Z, from a length-converted catch curve. The response is
log(N / Delta t), and the predictor is relative age from the inverse
von Bertalanffy growth curve. Z is estimated as the negative slope of
the regression.
}
\examples{
\dontrun{
data(alfonsino)
lfd <- lfd_alfonsino$Gillnet
lhpar <- lhpar_alfonsino

plot_zlcc(lfd, lhpar)

elc <- empirical_lc(lfd)
plot_zlcc(lfd, lhpar, lc = elc$lc)

out <- plot_zlcc(lfd, lhpar, return_data = TRUE)
out$fits
}

}
