% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccc.checks.R
\name{plot_zlcc_year}
\alias{plot_zlcc_year}
\title{Plot length-converted catch-curve Z by year}
\usage{
plot_zlcc_year(z, points = TRUE, line = TRUE, facet_iter = TRUE)
}
\arguments{
\item{z}{FLQuant returned by \code{\link[=zlcc]{zlcc()}} or data.frame returned by \code{\link[=zlcc_fits]{zlcc_fits()}}.}

\item{points}{Logical. Should points be drawn?}

\item{line}{Logical. Should lines be drawn?}

\item{facet_iter}{Logical. Should iterations be faceted when multiple iters
are present?}
}
\value{
A \code{ggplot} object.
}
\description{
Plots annual LCCC estimates of total mortality, Z. The input can either be an
FLQuant returned by \code{\link[=zlcc]{zlcc()}} or a data.frame returned by \code{\link[=zlcc_fits]{zlcc_fits()}}.
}
\examples{
\dontrun{
data(alfonsino)
lfd <- lfd_alfonsino$Gillnet
lhpar <- lhpar_alfonsino

z <- zlcc(lfd, lhpar)
plot_zlcc_year(z)
}

}
