% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcflicc.R
\name{prop_report}
\alias{prop_report}
\title{Propagate FLQuant objects in a FLicc report
f}
\usage{
prop_report(report, nsim)
}
\arguments{
\item{report}{A \code{fit$report} list from \code{fiticc()}.}

\item{nsim}{Number of Monte Carlo iterations.}
}
\value{
A report list where all \code{FLQuant} objects have an expanded \code{iter} dimension.
}
\description{
Propagate FLQuant objects in a FLicc report
f
}
