% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStockLen.R
\name{recl}
\alias{recl}
\title{Recruitment proxy from an FLStockLen object}
\usage{
recl(stk)
}
\arguments{
\item{stk}{An object of class \code{FLStockLen}.}
}
\value{
An \code{FLQuant} over year.
}
\description{
Returns the abundance in the first length class, used here as a recruitment
proxy for FLicc BRP objects.
}
