% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicccbrp.R
\name{refpts_flicc}
\alias{refpts_flicc}
\title{SPR-style reference points from an FLStockLen BRP object}
\usage{
refpts_flicc(stk, fit = 0, spr = c(40, 20, 10), labels = paste0("Fspr", spr))
}
\arguments{
\item{stk}{An object returned by \code{brp_stklen_flicc()}.}

\item{labels}{Character labels for the reference points.}

\item{sprx}{Numeric vector of SPR reference levels as proportions.}
}
\value{
An \code{FLPar} with SPR-based fishing reference points.
}
\description{
SPR-style reference points from an FLStockLen BRP object
}
