% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicc_mse.R
\name{spr_rule}
\alias{spr_rule}
\title{hcr_spr (flicc)}
\usage{
spr_rule(
  spr,
  b1 = 0.1,
  b2 = 0.35,
  b3 = 0.45,
  b4 = 0.6,
  dlow = -0.2,
  dopt = 0,
  dhi = 0.15,
  multiplier = TRUE
)
}
\arguments{
\item{spr}{Numeric SPR value(s).}

\item{b1}{Lower trigger point.}

\item{b2}{Trigger where no-change plateau starts.}

\item{b3}{Trigger where upper slope starts.}

\item{b4}{Trigger where maximum increase is reached.}

\item{dlow}{Relative change below b1, e.g. -0.20.}

\item{dopt}{Relative change between b2 and b3, usually 0.}

\item{dhi}{Relative change above b4, e.g. 0.10.}

\item{multiplier}{Logical, if TRUE return multiplier (default),
otherwise return relative change.}
}
\value{
Numeric vector of multipliers or relative changes.
}
\description{
SPR-based harvest control rule for FLicc.
}
\details{
Returns a multiplicative TAC/effort adjustment factor based on current SPR.
The rule follows the same piecewise structure as plot_hcrflicc():
\itemize{
\item spr <= b1: dlow
\item b1 < spr < b2: linear slope from dlow to dopt
\item b2 <= spr <= b3: dopt
\item b3 < spr < b4: linear slope from dopt to dhi
\item spr >= b4: dhi
}

Relative changes are converted to multipliers as:
multiplier = 1 + change
}
\examples{
# example code

spr <- seq(0,1,0.02)
change <- spr_rule(spr=seq(0,1,0.02),
b1 = 0.1,
b2 = 0.35,
b3 = 0.45,
b4 = 0.6,
dlow = -0.20,
dopt =  0.00,
dhi  =  0.15)
plot(spr,change-1,type="l",col=4,lwd=2,ylab=("Change TAC"),xlab="SPR")
abline(0,0,lty=2)
plot(spr,change,type="l",col=4,lwd=2,ylab=("Multiplier"),xlab="SPR")
abline(0,0,lty=2)

}
