% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStockLen.R
\name{ssbl}
\alias{ssbl}
\title{Spawning biomass by BRP year from an FLStockLen BRP object}
\usage{
ssbl(stk, spawn_time = attr(stk, "spawn_time", exact = TRUE) \%||\% 0)
}
\arguments{
\item{stk}{An object returned by \code{brp_stklen_flicc()}.}

\item{spawn_time}{Fraction of the interval elapsed before spawning.}
}
\value{
An \code{FLQuant} over the BRP year dimension.
}
\description{
Spawning biomass by BRP year from an FLStockLen BRP object
}
