% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fliccit.R
\name{ypr_flicc}
\alias{ypr_flicc}
\title{Yield-per-recruit from an FLicc fit}
\usage{
ypr_flicc(
  fit,
  nyears = 1,
  F = NULL,
  FM = NULL,
  Sel = NULL,
  scale_sel = TRUE,
  spawn_time = 0
)
}
\arguments{
\item{fit}{A fitted \code{"flicc_tmb_fit"} object.}

\item{nyears}{Integer. Number of terminal years to average over.}

\item{F}{Optional numeric scalar apical fishing mortality.}

\item{FM}{Optional numeric scalar apical \code{F/M} multiplier.}

\item{Sel}{Optional selectivity override.}

\item{scale_sel}{Logical. If \code{TRUE}, selectivity is scaled to an
apical maximum of 1.}

\item{spawn_time}{Numeric fraction of the year at which spawning occurs.}
}
\value{
Numeric YPR.
}
\description{
Yield-per-recruit from an FLicc fit
}
