% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flicc_plots.R
\name{plot_spr}
\alias{plot_spr}
\title{Plot spawning potential ratio by year}
\usage{
plot_spr(
  fit,
  title = NULL,
  by_year = 2,
  line_colour = "black",
  point_colour = "black",
  linewidth = 0.6,
  point_size = 2,
  ymax = NULL,
  reflines = TRUE,
  ref_levels = c(0.4, 0.2, 0.1),
  ref_colours = c("darkgreen", "orange", "red"),
  ref_linetype = 2
)
}
\arguments{
\item{fit}{A fitted \code{flicc_tmb_fit} object, list of fits or \emph{FLQuants}.}

\item{title}{Optional plot title.}

\item{line_colour}{Colour of the SPR time-series line.}

\item{point_colour}{Colour of the SPR points.}

\item{linewidth}{Line width for the SPR time-series line.}

\item{point_size}{Point size for the SPR observations.}

\item{ymax}{Optional upper y-axis limit. If \code{NULL}, this is set to the
maximum of the data, the reference levels, and \code{0.6}.}

\item{reflines}{Logical. If \code{TRUE}, horizontal reference lines are added.}

\item{ref_levels}{Numeric vector of SPR reference levels.}

\item{ref_colours}{Character vector of colours for the SPR reference lines.}

\item{ref_linetype}{Line type for the SPR reference lines.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots spawning potential ratio (SPR) by year from a fitted
\code{flicc_tmb_fit} object using \code{ggplot2}.
}
\details{
SPR values are taken from \code{fit$report$spr}, which is expected to be an
\code{FLQuant} indexed by year.

Reference lines can be toggled on or off using \code{reflines}. Their levels and
colours can be controlled with \code{ref_levels} and \code{ref_colours}.

The function expects the fitted object to contain:
\describe{
\item{\code{fit$report$spr}}{An \code{FLQuant} of spawning potential ratio by year.}
}
}
\examples{
\dontrun{

fit <- fiticc(lfd_alfonsino,stklen_alfonsino,sel_fun=c("dsnormal","logistic"),catch_by_gear = c(0.7,0.3))
plot_spr(fit)
plot_spr(fit, reflines = FALSE)
plot_spr(fit, ref_levels = c(0.7, 0.4, 0.2))
}

}
\seealso{
\code{\link[=plot_len]{plot_len()}}
}
