% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStockLen.R
\name{ssbl}
\alias{ssbl}
\title{Spawning stock biomass for FLStockLen}
\usage{
ssbl(stk)
}
\arguments{
\item{stk}{An \code{FLStockLen} object.}
}
\value{
An \code{FLQuant}.
}
\description{
Spawning stock biomass for FLStockLen
}
